Refactor WebSocket utilities and `useWebsocket` hook: modularize logic, integrate `useIsLoggedIn`, enhance state management, and remove unused `websockets.js`.
