+++ frontend/src/hooks/useIsLoggedIn.js
@@ -1,0 +1,6 @@
+import { useAuth } from './useAuth';
+
+export const useIsLoggedIn = () => {
+  const auth = useAuth();
+  return !!auth?.isLoggedIn;
+};
